+++ tests/docs/site/_quarto.yml
@@ -1,0 +1,17 @@
+project:
+  type: website
+
+website:
+  title: "."
+  navbar:
+    left:
+      - href: index.qmd
+        text: Home
+      - about.qmd
+      - href: subfolder/index.qmd
+
+format:
+  html:
+    theme: cosmo
+    css: styles.css
+    toc: true
