+++ tests/docs/site/about.qmd
@@ -1,0 +1,5 @@
+---
+title: "About"
+---
+
+About this site
