+++ tests/docs/site/styles.css
@@ -1,0 +1,1 @@
+/* css styles */
