+++ tests/docs/site/subfolder/index.qmd
@@ -1,0 +1,9 @@
+---
+title: "Cool Subfolder"
+---
+
+This is a Quarto website.
+
+To learn more about Quarto websites visit <https://quarto.org/docs/websites>.
+
+{{< embed ../notebooks/computations.ipynb#fig-visualization >}}
